requirements.txt
discord.py["voice"]
charset-normalizer
cryptography
pyyaml
aiohttp-client-cache
orjson
reactionmenu
sqids
mcstatus
aiohttp
discord-ext-help
pillow
numpy
jishaku
motor
aiosqlite
captcha
ulid